+++ scripts/scripts.js
@@ -244,6 +244,52 @@ export function getQueryParameter() {
   return params;
 }
 
+/**
+ * Set a cookie
+ * @param cname the name of the cookie
+ * @param cvalue the value of the cookie
+ * @param exdays the expiration days of a cookie
+ */
+export function setCookie(cname, cvalue, exdays) {
+  const date = new Date();
+  let hostName = '';
+  let domain = '';
+  let expires = '';
+  date.setTime(date.getTime() + (exdays * 24 * 60 * 60 * 1000));
+  if (exdays !== 0) {
+    expires = `expires=${date.toUTCString()}`;
+  }
+
+  domain = window.location.hostname.endsWith('.moleculardevices.com');
+  if (domain === true) {
+    hostName = 'domain=.moleculardevices.com;';
+  }
+  /* eslint-disable no-trailing-spaces */
+  document.cookie = `${cname}=${cvalue};secure;${hostName}${expires};path=/`;
+}
+
+/**
+ * Get a cookie
+ * @param cname the name of the cookie
+ */
+export function getCookie(cname) {
+  /* eslint-disable no-param-reassign */
+  cname += '=';
+  const decodedCookie = decodeURIComponent(document.cookie);
+  const ca = decodedCookie.split(';');
+  /* eslint-disable-next-line no-plusplus */
+  for (let i = 0; i < ca.length; i++) {
+    let c = ca[i];
+    while (c.charAt(0) === ' ') {
+      c = c.substring(1);
+    }
+    if (c.indexOf(cname) === 0) {
+      return c.substring(cname.length, c.length);
+    }
+  }
+  return '';
+}
+
 async function loadPage() {
   await loadEager(document);
   await loadLazy(document);
