+++ test.html
@@ -1,0 +1,17 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+  <meta charset="UTF-8">
+  <meta http-equiv="X-UA-Compatible" content="IE=edge">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <title>Document</title>
+</head>
+
+<body>
+ <iframe src="http://localhost:3000/cp-request?STYXKEY_PortalUserRole=dev">
+
+ </iframe>
+</body>
+
+</html>
